Update Delete's signature to (deleted, error)
If deleted is true, then the entry was deleted. Otherwise, the entry's
deletion is in-progress. The latter's useful when deleting e.g. VMs
because it gives us a way to let the user know that the VM's being deleted
without having to wait for the deletion, which can take a long time
(several minutes in GCP, and up to an hour in AWS).

Signed-off-by: Enis Inan <enis.inan@puppet.com>