Omit simple method invocation from history
Previously only activity that results in logged messages was added to
the history. When recording method invocation was added it adds
everything that triggers a method to history as a side-effect. That's
not necessary and results in operations that simply stat a file being
included. Remove adding them to history.

Signed-off-by: Michael Smith <michael.smith@puppet.com>