Download busybox if not present for listing Docker volumes
The Docker plugin uses a busybox image to explore persistent volumes by
spinning up an instance that mounts the volume and then interacting with
it. If the busybox image is not present, creating the instance will fail
rather than pull the image. Explicitly pull the busybox:latest image if
it's missing.

Tested as part of Travis CI. We previously would explicitly pull the
busybox image so it was present when running `wash validate` on the
Docker plugin.

Note that this isn't a problem for kubernetes as kubernetes already
pulls images on-demand.

Fixes #312.

Signed-off-by: Michael Smith <michael.smith@puppet.com>