Increase scanner buffer size for whistory
bufio.Scanner has a default buffer size of 64KiB. If a line is longer
than that it stops scanning and errors. I encountered this problem with
a long line from printing the output of listing a large directory, and
whistory hid the error.

Ensure the error is returned if scanning fails. Also increase the buffer
size to 100MiB so it's unlikely we'll encounter this error.

Fixes #580.

Signed-off-by: Michael Smith <michael.smith@puppet.com>