Generalize console control
Sometimes a plugin needs control of the console for a prompt it doesn't
make directly. Get control before calling all primitives that take a
context (as a proxy for potential network activity or execs).

This isn't entirely safe. If it's backgrounded and prompts for input it
may not exit cleanly. Using `plugin.Prompt` guards against this issue.

Fixes #297.

Signed-off-by: Michael Smith <michael.smith@puppet.com>