Support multiple paths for signal
This also changes the usage from `signal <path> <signal>` to `signal
<signal> <path>` (which is what we had in a previous commit). The latter
usage generalizes well to multiple paths; the former usage requires a
separate `--path` option which is more complexity for very little gain.

We can still have a clean usage for signal-specific options with `signal
<signal> <path>`. It would probably look something like:

* `signal --help` -- Print signal's usage

* `signal --help <path>` -- List the supported signals/signal groups of
path

* `signal <signal> --help <path>` -- List the CLI opts associated with
<signal> and <path>

which is not that bad.

Signed-off-by: Enis Inan <enis.inan@puppet.com>