Make Wash daemon part of the shell session
When `wash` is run, it starts a shell as a child process where all the
action happens. It then stays running as a daemon to handle
Wash-specific requests through FUSE and the socket API. However if it
needs to prompt for input, it has trouble because it's TTY is not the
same as the one handling input (the child shell).

We previously addressed that by temporarily transferring control of
input to the Wash daemon, but that had some problems
- it only worked when a plugin requested input using Wash's helper
- if the process that triggered the plugin activity tried to do
something with stdin - like MacVim appears to on startup - it would fail

Reparent the Wash daemon to have a new session under the new shell so
they share the same TTY. Prompts for input now work wherever they're
triggered from. MacVim still behaves a little strangely if you get a
prompt when opening it, but I'm not sure what to do about it.

Fixes #623.

Signed-off-by: Michael Smith <michael.smith@puppet.com>