Ensure the original process group is preserved for zsh
If Wash starts as the leader of its own process group, then when we move
the Wash daemon to a new session we change the permission model of that
original process group.  Some shells, zsh in particular, try to return
the tty to the parent process group when they exit. When that happens
and the parent's process group is now under in a different session, this
fails with
```
zsh: can't set tty pgrp: operation not permitted
```

When Wash detects it starts as the leader of its process group, fork so
that the daemon is not the leader so we create a new process group when
reparenting it under the shell.

Signed-off-by: Michael Smith <michael.smith@puppet.com>