Address feedback and fix issues around tracking file size
Add `sizeValid` to track whether we need to call `plugin.Size` before
using `size`. This is only needed when opening a file WriteOnly to delay
a Read - to ensure we have the entire file content before Write - so we
only do it when necessary.

Handle `io.EOF` when loading the file in case a call to `Setattr`
increased it beyond the original size but didn't write to fill that
space, and pad with null characters when that happens.

Upgrade journal entries to warnings when they result in an error. FUSE
doesn't relay the actual error messages.

Delay releasing data and clearing entry cache until we release all
writers.

Tweak a few tests so we have a balance of ReadWrite tests that rely on
attribute size or calling Read to determine size. Update
TestTruncateAndWrite to ensure we don't call Read when we've called
Setattr to truncate the file first.

Add additional comments explaining how to use BlockReadable and
Writable, and some code comments to clarify how Open works.

Signed-off-by: Michael Smith <michael.smith@puppet.com>