Move the external plugin implementation to its own package
Separates the external plugin implementation into a separate package
from `plugin` so that it can make use of other packages in Wash that
themselves depend on `plugin` (like `volume` and `transport`).

The separation is possible by making `MethodSignature` a public type,
and adding a hidden `externalPlugin` interface for "dynamic" plugin
functionality ( where an entry's features may not be known at
compile-time).

Signed-off-by: Michael Smith <michael.smith@puppet.com>