Add the initial rql package layout and the predicate interfaces
The separate Eval* methods will let us come up with a generic predicate
expression type while also giving us compile-time type-checking where it
is needed (the compile-time type-checking ensures that something like
the Action primary doesn't accidentally get a numeric predicate). The
*InDomain methods let us properly handle negation. Both these changes result
in a simpler predicate implementation than the previous `wash find` code because
we no longer have to create custom And/Or/Not operator classes.

Note that the previous `wash find` negation code was complicated because the
predicate interface design was conflating the "check if this predicate makes
sense for this kind of entry/metadata value" [domain] and "evaluate
the predicate for this OK value" [the corresponding Eval* method]) parts
of an entry/value predicate.

Signed-off-by: Enis Inan <enis.inan@puppet.com>