Add the generic predicate expression type
Note that all subsequent predicates and primaries will be wrapped in the
Atom type. The Atom type's a useful place to centralize the
EntryPredicate, EntrySchemaPredicate, and ValuePredicate semantics
mentioned in rql/interfaces.go. Also, the primary predicate expression
code is in expression.go.

Some of the tests use predicates that haven't yet been shown. These
tests are meant to test And/Or's truth tables for the appropriate Eval*
methods; hopefully that part is clear from the code.

Subsequent commits will include the separate predicate/primary
implementations. These implementations will have a corresponding
TestExpression_AtomAndNot test case that ensures that they can be
marshaled as atoms, that they are negated correctly and they will likely
work with the general predicate expression interface.

Signed-off-by: Enis Inan <enis.inan@puppet.com>