Remove the *InDomain methods
The same thing could be achieved with proper documentation. E.g. We
could encourage people to _not_ negate primaries of the form
[<primary_name>, <predicate>] as ["NOT", [<primary_name>, <predicate>]]
but instead write it as [<primary_name>, ["NOT", <predicate>]]. So for
example, something like ["NOT", ["path", ["glob", "foo"]]] could be
written as ["path", ["NOT", ["glob", "foo"]]]. The latter still negates
the path _without_ including the start path (which the former would do).

Similarly for value predicates, something like ["NOT", ["number", [">",
"5"]]] could be written as ["number", ["NOT", [">", "5"]]]. This would
still negate the predicate while also returning false for mis-typed
values.

Hence, the *InDomain methods are not necessary. This considerably
simplifies the code, as will be seen in later commits.

Signed-off-by: Enis Inan <enis.inan@puppet.com>