Improve reporting on non-zero exit for fs List
Some non-zero exits are normal with `find` (see examples in code). The
messages that caused this can appear on stdout or stderr depending on
whether a tty is used or if it's a `stat` executed by `find`. Ignore
normal failures, but if we get an unknown one fail `VolumeList` and
return the error.

Fixes #742.

Signed-off-by: Michael Smith <michael.smith@puppet.com>