Cleanup API docs
The last argument for a `swagger:route` is a long ID, not the response
type. The docs generator also expects a summary string to put as the
description for the method. Update documentation markup to match that
and remove the redundant `findResponse` (`entryList` is already generic
enough and was already being used for the 200 response type).

Also switch to only defining the parameters for an endpoint in one place
(for `exec`).

Signed-off-by: Michael Smith <michael.smith@puppet.com>