Inspect volume on existing pod if already mounted
This speeds up access and avoids errors trying to mount a volume in a
new pod when it can only be mounted once.

Creates the containerBase type to represent a specific Kubernetes
container and provide methods to execute a command and asynchronously
stream its output. This reduces some of the duplication across container
and persistent volume claim implementations.

Simplify execution logic by using the same method on an existing
container or when we create a new temporaryContainer.

Fixes #758.

Signed-off-by: Michael Smith <michael.smith@puppet.com>